+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.6 - Sentence split
+
+* Split sentences and join but not longer than 450 characters
+* No more one letter sentences
+
 ## 0.0.5 - Lower case
 
 * Let users use mixed case